search/crawler/requirements.txt
certifi==2024.6.2
elastic-transport==8.13.1
elasticsearch==8.14.0
redispy==3.0.0
urllib3==2.2.2
